(Re)Enable the play button on the season page
Resurrecting some of the previously-commented code to enable the play
button on the season page. Which plays the next unplayed episode in that
season. Note that we no longer bail out on the `BaseItemDto.missing`
flag because sometimes it shows false negatives even on legit season
items. Instead, we rely on results returning from the API call.